Remove directories
I also removed the client directory because I can't manage to get that client to display things from the server. I mean, that's part of why we wanted to update the lab -- it's a different way to interact with the data (expects flat files). It might be able to come back in, but it would need some love, and I think it will be better to remove it until it's working how we want.