Update to clarify structure of the lab
The previous commits on this branch did a nice job getting the creation of GitHub Projects better documented for the lab. This commit focuses on making sure that the students will be able to tell at a glance how the tasks are structured. In particular, I want them to notice that exploring the project, answering questions, and setting up the GitHub Projects aspects of the lab should happen before the coding parts of the lab. I also wanted to draw their attention to how they will be graded (without putting those details here, since it might change over time and changing it in two places is no fun). Finally, I updated some wording (e.g., not using _issues_ except where we mean GitHub issues)

I added a couple of headings to the coding part of the lab as well because I think that students will be confused if almost all of the things in the table of contents at the top are not the coding parts of the lab. This made it a bit more balanced.