Add numbering to checkbox lists in a consistent way
The numbering doesn't render the way it was written before, so this commit moves the numbering of checkbox items to be after the checkbox. I added numbering to the sub lists to follow the format of the existing list about making a milestone.

There are lots of checkboxes in this lab, and I think the new documentation about making a project just highlighted that not only are there lots of checklists here, but there are some inconsistencies in how and why checkboxes may be used.

I don't think this is perfect, but it is (more) consistent throughout the lab now. I like and am used to having numbered tasks. In data structures, I tend to use //TODO for graded elements. It's messier in this lab what it means for something to be a graded item since we _are_ grading things like the use of GitHub Projects. I think that using checkboxes makes it a little more obvious that these are tasks the students should be doing, and that they should be doing those setup tasks _before_ they begin the coding parts of the lab. It's very tempting to just skip down to the coding part. I think the checkboxes may help reduce that temptation.